cpp-devel-txt
Finished adding in global and local command lists. The local commands are
essentially a list of functions at this moment but I will be adding text
to each entry as I work my way through the file/functions.

I have also added a section to each GlobalCommand entry showing the
command to remap in the users .vimrc file.