cpp-devel-vim.vim:
Removing old commented out lines and options. These were unused options
and the original plan to store this information in this file is
pointless and useless because no one will actually open the plugin
script to add information about compilers. -i.e. having this information
in this file could cause major problems later if updating to new
compilers or methods (a programmer has to maintain a Makefile or
CMakeLists file, and that should be the driving force, not a helper
script).